requirements.txt
azure-storage-blob
azure-storage
cryptography==2.5
pyyaml